requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
arrow==1.2.3
async-timeout==4.0.2
attrs==22.2.0
binaryornot==0.4.4
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
cookiecutter==2.1.1
distlib==0.3.6
django-admin-geomap==1.3
filelock==3.8.0
Flask==2.2.3
frozenlist==1.3.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jinja2-time==0.2.0
MarkupSafe==2.1.2
multidict==6.0.4
openai==0.27.2
openai-secret-manager==0.0.2
platformdirs==2.5.4
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
python-telegram-bot==20.2
PyYAML==6.0
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
text-unidecode==1.3
tqdm==4.65.0
urllib3==1.26.14
virtualenv==20.17.0
Werkzeug==2.2.3
yarl==1.8.2
